Split assets store
