+++ src/store/assets/detail/index.js
@@ -1,0 +1,18 @@
+import actions from './actions';
+import getters from './getters';
+import mutations from './mutations';
+
+export const initState = () => ({
+  error: null,
+  isLoading: false,
+  isHandling: false,
+  detail: null,
+});
+
+export default {
+  namespaced: true,
+  state: initState(),
+  getters,
+  mutations,
+  actions,
+};
