+++ src/store/assets/ibc/getters.js
@@ -1,0 +1,7 @@
+export default {
+  error: (state) => state.error,
+  isLoading: (state) => state.isLoading,
+  isHandling: (state) => state.isHandling,
+  connection: (state) => state.connection,
+  modal: (state) => state.modal,
+};
