+++ src/store/assets/ibc/index.js
@@ -5,17 +5,9 @@ import mutations from './mutations';
 export const initState = () => ({
   error: null,
   isLoading: false,
-  isFetching: false,
   isHandling: false,
-  isInvalid: true,
   connection: null,
-  supply: [],
-  list: [],
-  balances: null,
   modal: null,
-  detail: null,
-  newCW20: null,
-  government: '',
 });
 
 export default {
