+++ src/store/assets/list/index.js
@@ -1,0 +1,19 @@
+import actions from './actions';
+import getters from './getters';
+import mutations from './mutations';
+
+export const initState = () => ({
+  error: null,
+  isLoading: false,
+  balances: null,
+  list: [],
+  supply: [],
+});
+
+export default {
+  namespaced: true,
+  state: initState(),
+  getters,
+  mutations,
+  actions,
+};
