+++ src/store/assets/new/index.js
@@ -1,0 +1,20 @@
+import actions from './actions';
+import getters from './getters';
+import mutations from './mutations';
+
+export const initState = () => ({
+  error: null,
+  isLoading: false,
+  isHandling: false,
+  isInvalid: true,
+  government: '',
+  newCW20: null,
+});
+
+export default {
+  namespaced: true,
+  state: initState(),
+  getters,
+  mutations,
+  actions,
+};
