+++ src/views/assets/detail/marketing/AssetsModalLogo.vue
@@ -59,7 +59,7 @@ export default {
     };
   },
   computed: {
-    ...mapGetters('assets', ['isHandling']),
+    ...mapGetters('assetsDetail', ['isHandling']),
     dialog: {
       get() {
         return this.value;
