+++ src/views/assets/detail/marketing/AssetsModalMarketing.vue
@@ -79,7 +79,7 @@ export default {
     };
   },
   computed: {
-    ...mapGetters('assets', ['isHandling']),
+    ...mapGetters('assetsDetail', ['isHandling']),
     dialog: {
       get() {
         return this.value;
