+++ src/views/assets/list/AssetsModalTransfer.vue
@@ -1,6 +1,6 @@
 <template>
   <v-dialog max-width="800" transition="dialog-transition" v-model="dialog">
-    <v-card flat :loading="isFetching">
+    <v-card flat :loading="isLoading">
       <v-toolbar class="mb-3" flat>
         <v-toolbar-title class="primary--text">
           <span v-text="title" />
@@ -80,7 +80,7 @@ export default {
     };
   },
   computed: {
-    ...mapGetters('assets', ['isFetching', 'isHandling', 'modal']),
+    ...mapGetters('assetsIbc', ['isLoading', 'isHandling', 'modal']),
     balance() {
       return this.token ? this.token.balance : '0';
     },
@@ -120,7 +120,7 @@ export default {
     this.chain = this.connections[0];
   },
   methods: {
-    ...mapActions('assets', ['handleModal', 'initIBCTransfer']),
+    ...mapActions('assetsIbc', ['handleModal', 'initIBCTransfer']),
     close() {
       this.handleModal(null);
     },
